+++ src/screens/account_details/components/index.ts
@@ -1,11 +1,11 @@
 import Profile from './profile';
 import Transactions from './transactions';
 import Overview from './overview';
-import OtherTokens from './other_tokens';
+import Tokens from './tokens';
 
 export {
   Profile,
   Transactions,
   Overview,
-  OtherTokens,
+  Tokens,
 };