+++ src/screens/account_details/components/tokens/components/list/components/desktop/index.tsx
@@ -1,62 +1,38 @@
 import React from 'react';
 import classnames from 'classnames';
 import useTranslation from 'next-translate/useTranslation';
-import Link from 'next/link';
 import {
   Table,
   TableHead,
   TableRow,
   TableCell,
   TableBody,
-  Typography,
 } from '@material-ui/core';
-import { getMiddleEllipsis } from '@utils/get_middle_ellipsis';
-import { TRANSACTION_DETAILS } from '@utils/go_to_page';
-import {
-  Result, AvatarName,
-} from '@components';
-import dayjs from '@utils/dayjs';
+import { formatNumber } from '@utils/format_token';
+import { TOKEN_DETAILS } from '@utils/go_to_page';
+import { AvatarName } from '@components';
 import { columns } from './utils';
 import { useStyles } from './styles';
-import { Shard } from '..';
+import { OtherTokenType } from '../../../../types';
 
-const Desktop: React.FC<{items: TransactionType[]} & ComponentDefault> = (props) => {
-  const { t } = useTranslation('transactions');
+const Desktop: React.FC<{items: OtherTokenType[]} & ComponentDefault> = (props) => {
+  const { t } = useTranslation('accounts');
   const classes = useStyles();
   const formattedItems = props.items.map((x) => {
     return ({
-      hash: (
-        <Link href={TRANSACTION_DETAILS(x.hash)} passHref>
-          <Typography variant="body1" className="value" component="a">
-            {getMiddleEllipsis(x.hash, {
-              beginning: 10, ending: 10,
-            })}
-          </Typography>
-        </Link>
-      ),
-      shard: <Shard to={x.toShard} from={x.fromShard} />,
-      from: (
+      identifier: x.identifier,
+      token: (
         <AvatarName
-          address={x.from}
-          name={getMiddleEllipsis(x.from, {
-            beginning: 10, ending: 10,
-          })}
+          address={x.identifier}
+          imageUrl={x.imageUrl}
+          name={x.name}
+          href={TOKEN_DETAILS}
         />
       ),
-      to: (
-        <AvatarName
-          address={x.to}
-          name={getMiddleEllipsis(x.to, {
-            beginning: 10, ending: 10,
-          })}
-        />
-      ),
-      status: (
-        <Result status={x.status} />
-      ),
-      time: dayjs.utc(dayjs.unix(x.timestamp)).fromNow(),
+      balance: `${formatNumber(x.balance.value, x.balance.exponent)} ${x.balance.displayDenom.toUpperCase()}`,
     });
   });
+
   return (
     <div className={classnames(props.className, classes.root)}>
       <Table>