+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 - Add Transactions to Validator Details if provider
 - Add Tokens Page
 - Add Tokens Details Page
+- Add Tokens to Account Details